Move `Spree::Event::MailerProcessor under `Spree::Event::Porcessors` namespace
Since there can be multiple event processors, it makes sense having a dedicated
namespace for them.
