+++ core/spec/models/spree/order_spec.rb
@@ -31,11 +31,11 @@ RSpec.describe Spree::Order, type: :model do
       # all the ones set by MailerProcessor module
       context 'when removing the default email notification subscription' do
         before do
-          Spree::Event.unsubscribe Spree::Event::MailerProcessor.order_finalize_subscription
+          Spree::Event.unsubscribe Spree::Event::Processors::MailerProcessor.order_finalize_subscription
         end
 
         after do
-          Spree::Event::MailerProcessor.order_finalize
+          Spree::Event::Processors::MailerProcessor.order_finalize
         end
 
         it 'does not send the email' do
@@ -46,11 +46,11 @@ RSpec.describe Spree::Order, type: :model do
 
       context 'when removing all the email notification subscriptions' do
         before do
-          Spree::Event::MailerProcessor.unregister!
+          Spree::Event::Processors::MailerProcessor.unregister!
         end
 
         after do
-          Spree::Event::MailerProcessor.register!
+          Spree::Event::Processors::MailerProcessor.register!
         end
 
         it 'does not send the email' do
